+++ spec/Storefront/Controller/ResultControllerSpec.php
@@ -81,7 +81,7 @@ class ResultControllerSpec extends ObjectBehavior
     public function it_should_return_original_response_content_when_ssr_is_not_active(
         SearchAdapter $searchAdapter,
         Page $page,
-        Engine $mustache
+        Engine $handlebars
     ): void {
         $content = 'original content';
         $this->pageLoader->load($this->request, $this->salesChannelContext)->willReturn($page);
@@ -93,7 +93,7 @@ class ResultControllerSpec extends ObjectBehavior
             $this->request,
             $this->salesChannelContext,
             $searchAdapter,
-            $mustache
+            $handlebars
         );
 
         $response->shouldBeAnInstanceOf(Response::class);